requirements.txt
python-fasthtml
monsterui
python-dotenv
fastmigrate
cdp-sdk
